Add first draft of master-slave multiblock system
Add IMultiblock which is a way to organise a multiblock structure and its rules without binding it to a block or tile entity.

Add Master, Slave and Type as ways to organise the multiblock structure.

Add an example which needs to be implemented fully for any flaws to be ironed out.

Add ITask and Executor which are what the different parts of the multiblock will be capable of doing.

Still required: Link slaves with tasks, and executors so the work can be delegated to different blocks or pooled into the master depending on the requirements of each multiblock.